Move State to data package
