+++ core/src/main/scala/cats/data/StateT.scala
@@ -1,7 +1,5 @@
 package cats
-package state
-
-import cats.data.Kleisli
+package data
 
 /**
  * `StateT[F, S, A]` is similar to `Kleisli[F, S, A]` in that it takes an `S`
@@ -123,7 +121,7 @@ object StateT extends StateTInstances {
     StateT(s => F.pure((s, a)))
 }
 
-private[state] sealed abstract class StateTInstances {
+private[data] sealed abstract class StateTInstances {
   implicit def stateTMonadState[F[_], S](implicit F: Monad[F]): MonadState[StateT[F, S, ?], S] =
     new MonadState[StateT[F, S, ?], S] {
       def pure[A](a: A): StateT[F, S, A] =
@@ -143,7 +141,7 @@ private[state] sealed abstract class StateTInstances {
 
 // To workaround SI-7139 `object State` needs to be defined inside the package object
 // together with the type alias.
-private[state] abstract class StateFunctions {
+private[data] abstract class StateFunctions {
 
   def apply[S, A](f: S => (S, A)): State[S, A] =
     StateT.applyF(Now((s: S) => Now(f(s))))
