--- core/src/main/scala/cats/state/package.scala
@@ -1,6 +1,0 @@
-package cats
-
-package object state {
-  type State[S, A] = StateT[Eval, S, A]
-  object State extends StateFunctions
-}
