+++ tests/src/test/scala/cats/tests/FreeApplicativeTests.scala
@@ -5,8 +5,7 @@ import cats.arrow.NaturalTransformation
 import cats.free.FreeApplicative
 import cats.laws.discipline.{CartesianTests, ApplicativeTests, SerializableTests}
 import cats.laws.discipline.eq.{tuple3Eq, tuple2Eq}
-import cats.data.Const
-import cats.state.State
+import cats.data.{Const, State}
 
 import org.scalacheck.{Arbitrary, Gen}
