+++ tests/src/test/scala/cats/tests/StateTTests.scala
@@ -2,7 +2,7 @@ package cats
 package tests
 
 import cats.laws.discipline.{CartesianTests, MonadStateTests, MonoidKTests, SerializableTests}
-import cats.state.{State, StateT}
+import cats.data.{State, StateT}
 import cats.laws.discipline.eq._
 import cats.laws.discipline.arbitrary._
 import org.scalacheck.{Arbitrary, Gen}
